add_battons

+++ bot_utils.py
@@ -2,33 +2,47 @@ from datetime import datetime
 from ftplib import FTP
 import os
 import re
+from adm_menu import admin_menu
 import cf
 import locale
-from mysql.connector import connect, Error
+from adm_menu import admin_menu
 
 # настройки подключеня к серверам по FTP
 config_m_1 = [cf.M1_IP, cf.M1_PORT, cf.M1_LOGIN, cf.M1_PASS]
 config_m_3 = [cf.M3_IP, cf.M3_PORT, cf.M3_LOGIN, cf.M3_PASS]
 
+# проверяем имя сервера с которым будем работать
+def path_from_server(name_server, path_m1, path_m3):
+    if name_server == 'm1':
+        return path_m1
+    elif name_server == 'm3':
+        return path_m3
+    write_to_log(datetime.now().strftime("%d-%m-%Y %H:%M") + 'Error server name ', name_server)
+    return 0
+
+
 def find_actual_file() -> str:
     max_date = ['0'] * 3
     result = ''
     date = datetime.now().strftime("%m-%d-%Y")
-
-    with open("listFiles", 'r', encoding='utf-8') as file:
-        list_files = file.readlines()
-        for line in list_files:
-            if line.find(date) >= 0 and line.find('bin') >= 0 :
-                result = line
-                break
-            str_date = line.split(' ', 1)
-            if re.fullmatch("\d\d-\d\d-\d{4}", str_date[0]):
-                tmp = str_date[0].split('-')
-                if int(tmp[2]) > int(max_date[2]) or int(tmp[0]) > int(max_date[0]) or int(tmp[1]) > int(max_date[1]):
-                    max_date = tmp
+    try:
+        with open("listFiles", 'r', encoding='utf-8') as file:
+            list_files = file.readlines()
+            for line in list_files:
+                if line.find(date) >= 0 and line.find('bin') >= 0 :
                     result = line
-    print(result)
-    os.remove("listFiles")
+                    break
+                str_date = line.split(' ', 1)
+                if re.fullmatch("\d\d-\d\d-\d{4}", str_date[0]):
+                    tmp = str_date[0].split('-')
+                    if int(tmp[2]) > int(max_date[2]) or int(tmp[0]) > int(max_date[0]) or int(tmp[1]) > int(max_date[1]):
+                        max_date = tmp
+                        result = line
+    except:
+        print('Error read listFiles.')
+        return result
+    if os.path.exists('listFiles'):
+        os.remove("listFiles")
     return result
 
 # запись логов в файл
@@ -43,21 +57,16 @@ def write_to_log(str, server):
 
 # функция для подключения к FTP
 def connect_snavi_ftp(serv):
-    ftp = FTP()
-    if serv == 'm1':    # проверяем валидное ли имя сервера
-        s = config_m_1
-    elif serv == 'm3':
-        s = config_m_3
-    else:
-        write_to_log('Error server name ', serv)
-        ftp.quit()
+    s = path_from_server(serv, config_m_1, config_m_3)
+    if s == 0:
         return 0
+    ftp = FTP()
     try:
-        print(ftp.connect(s[0], int(s[1]))) # подключаемся к ftp
+        print(ftp.connect(s[0], int(s[1])))
         print(ftp.login(s[2], s[3]))
     except:
-        write_to_log('Error connect to ftp ',s[0])
-        ftp.quit()
+        write_to_log('Error connect to ftp: ',s[0])
+        ftp.close()
         return 0
     return ftp
 
@@ -77,17 +86,12 @@ def find_num_in_string(line, number):
 def check_file(number, serv, file):
     date = datetime.now() 
     ftp = connect_snavi_ftp(serv)
-    if ftp == 0:
-        return 0
-    if serv == 'm1':
-        path = file
-    elif serv == 'm3':
-        path = ('AG5/' + file)
+    path = path_from_server(serv, file, 'AG5/'+ file)
     try:
         with open(cf.file_tmp, 'wb') as f_tmp:
-            ftp.retrbinary('RETR ' + path, f_tmp.write, 1024)
+            ftp.retrbinary('RETR ' + path, f_tmp.write)
     except:
-        write_to_log('Error file ' + path + ' write from ftp ', serv)
+        write_to_log('Error file: ' + path + ' write from ftp ', serv)
         ftp.quit()
         return 0
     with open(cf.file_tmp, 'r') as f:
@@ -103,19 +107,19 @@ def check_file(number, serv, file):
     ftp.quit()
     return 0
 
-    # создаем файл atg в чужих
+
+# создаем файл atg в чужих
 def add_to_atg(update, context):
     count = 0
     chat = update.effective_chat
     text = update.message.text.split()
-    now = datetime.now()
     locale.setlocale(locale.LC_ALL, '')
     file_atg = ['AutoGRAPH database 1.3\n', str(count), ' records\n']
-    path  = cf.path_predator + now.strftime("%m. %B\\")
+    path  = cf.path_predator + datetime.now().strftime("%m. %B\\")
     if not os.path.isdir(path):
         os.mkdir(path)
     if len(text) == 4:
-        if text[2].isdigit() and len(text[2]) == 7 and text[3].isdigit() and len(text[3]) == 7:
+        if re.fullmatch(r'\d{7}', text[2]) and re.fullmatch(r'\d{7}', text[3]):
             count = int(text[3]) - int(text[2]) + 1
             count1 = count
             file_atg[1] = str(count)
@@ -149,17 +153,29 @@ def add_to_atg(update, context):
     else:
         context.bot.send_message(chat_id=chat.id, text=chat.first_name + ', аргуентов должно быть три:\n/atg [имя файла] [номер прибра первый] [номер последний]\n Если нужно добавить в один и тот же файл еще приборы, просто повторите операцию.')
 
+
 #отправляем уведомление админу и добавляем юзера в базу
 def add_user(chat, context, id: str):
     context.bot.send_message(chat_id=id, text="⚠️У нас новый пользователь: " + chat.first_name)
-    with open('data', 'a', encoding='utf-8') as f:
-        f.write(str(chat.id) + ' ' + str(chat.first_name) + ' ' + str(chat.last_name) + '\n')
+    admin_menu(context, id, str(chat.id), chat.first_name)
+
+def parse_message_from_db(message)->str:
+    mess_list = ['🔸ID: ', '🔸№Прибора: ','🔸PASS: ', '🔸ТЕЛ1: ', '🔸ТЕЛ2: ','🔸Компания: ','🔸Дата: ', '🔸Файл: ', '🔸IMEI: ']
+    resault:str = ''
+    for line in message:
+        for (i, j) in zip(mess_list, line):
+            resault += i + str(j) + '\n'
+        resault += '#####################\n'
+    return resault
 
 
-#поиск днных в atg и добавление в базу данных
-def connect_to_db():
+# проверяем data файл на наличие id пользователя в нем
+def check_user(file, id):
     try:
-        with connect(host=cf.SQL_SERVER, user=cf.SQL_LOGIN, password=cf.SQL_PASS) as connection:
-            print(connection)
-    except Error as e:
-        print(e)
+        with open(file, 'r', encoding='utf-8') as f:
+            for line in f.readlines():
+                if re.match(str(id), line):
+                    return 1
+    except:
+        print(f'Ошибка чтения файла: {file}')
+    return 0
